requirements.txt
ibm-watsonx-ai==1.0.5
flask==2.3.3
python-dotenv==1.0.0